+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+0.7.3 (2019-02-10)
+-------------------
+- added dynamic load intensity adjustment. There are two new sections in config: 'redis' and 'delay_coefficient'.
+- moved class Timer from 'uploader' to 'app'.
+
 0.7.2 (2019-02-05)
 -------------------
 - added template SegmentsWithTimestamp.